+++ assets/js/dashboard/stats/graph/graph-util.js
@@ -1,5 +1,31 @@
-import { METRIC_LABELS, METRIC_FORMATTER } from './visitor-graph'
 import { parseUTCDate, formatMonthYYYY, formatDay } from '../../util/date'
+import numberFormatter, {durationFormatter} from '../../util/number-formatter'
+
+export const METRIC_MAPPING = {
+  'Unique visitors (last 30 min)': 'visitors',
+  'Pageviews (last 30 min)': 'pageviews',
+  'Unique visitors': 'visitors',
+  'Visit duration': 'visit_duration',
+  'Total pageviews': 'pageviews',
+  'Bounce rate': 'bounce_rate',
+  'Unique conversions': 'conversions',
+}
+
+export const METRIC_LABELS = {
+  'visitors': 'Visitors',
+  'pageviews': 'Pageviews',
+  'bounce_rate': 'Bounce Rate',
+  'visit_duration': 'Visit Duration',
+  'conversions': 'Converted Visitors',
+}
+
+export const METRIC_FORMATTER = {
+  'visitors': numberFormatter,
+  'pageviews': numberFormatter,
+  'bounce_rate': (number) => (`${number}%`),
+  'visit_duration': durationFormatter,
+  'conversions': numberFormatter,
+}
 
 export const dateFormatter = (interval, longForm) => {
   return function(isoDate, _index, _ticks) {
@@ -86,20 +112,6 @@ export const GraphTooltip = (graphData, metric) => {
       return !prev_label ? formattedLabel : prev_formattedLabel
     }
 
-    // function renderComparison(change) {
-    //   const formattedComparison = numberFormatter(Math.abs(change))
-
-    //   if (change > 0) {
-    //     return `<span class='text-green-500 font-bold'>${formattedComparison}%</span>`
-    //   }
-    //   if (change < 0) {
-    //     return `<span class='text-red-400 font-bold'>${formattedComparison}%</span>`
-    //   }
-    //   if (change === 0) {
-    //     return `<span class='font-bold'>0%</span>`
-    //   }
-    // }
-
     // Set Tooltip Body
     if (tooltipModel.body) {
       var bodyLines = tooltipModel.body.map(getBody);
@@ -113,28 +125,23 @@ export const GraphTooltip = (graphData, metric) => {
       const label = graphData.labels[data.dataIndex]
       const point = data.raw || 0
 
-      // const prev_data = tooltipModel.dataPoints.slice(-1)[0]
-      // const prev_label = graphData.prev_labels && graphData.prev_labels[prev_data.dataIndex]
-      // const prev_point = prev_data.raw || 0
-      // const pct_change = point === prev_point ? 0 : prev_point === 0 ? 100 : Math.round(((point - prev_point) / prev_point * 100).toFixed(1))
-
       let innerHtml = `
-			<div class='text-gray-100 flex flex-col'>
-				<div class='flex justify-between items-center'>
-						<span class='font-semibold mr-4 text-lg'>${METRIC_LABELS[metric]}</span>
-				</div>
-				<div class='flex flex-col'>
-					<div class='flex flex-row justify-between items-center'>
-						<span class='flex items-center mr-4'>
-							<div class='w-3 h-3 mr-1 rounded-full' style='background-color: rgba(101,116,205)'></div>
-							<span>${renderLabel(label)}</span>
-						</span>
-						<span class='text-base font-bold'>${METRIC_FORMATTER[metric](point)}</span>
-					</div>
-				</div>
-				<span class='font-semibold italic'>${graphData.interval === 'month' ? 'Click to view month' : graphData.interval === 'date' ? 'Click to view day' : ''}</span>
-			</div>
-			`;
+      <div class='text-gray-100 flex flex-col'>
+        <div class='flex justify-between items-center'>
+          <span class='font-semibold mr-4 text-lg'>${METRIC_LABELS[metric]}</span>
+        </div>
+        <div class='flex flex-col'>
+          <div class='flex flex-row justify-between items-center'>
+            <span class='flex items-center mr-4'>
+              <div class='w-3 h-3 mr-1 rounded-full' style='background-color: rgba(101,116,205)'></div>
+              <span>${renderLabel(label)}</span>
+            </span>
+            <span class='text-base font-bold'>${METRIC_FORMATTER[metric](point)}</span>
+          </div>
+        </div>
+        <span class='font-semibold italic'>${graphData.interval === 'month' ? 'Click to view month' : graphData.interval === 'date' ? 'Click to view day' : ''}</span>
+      </div>
+      `;
 
       tooltipEl.innerHTML = innerHtml;
     }
@@ -201,7 +208,6 @@ export const buildDataSet = (plot, present_index, ctx, label, isPrevious) => {
       label,
       data: plot,
       borderWidth: 2,
-      // borderDash: [10, 1],
       borderColor: 'rgba(166,187,210,0.5)',
       pointHoverBackgroundColor: 'rgba(166,187,210,0.8)',
       pointBorderColor: 'transparent',
